+++ src/main/java/net/coreprotect/database/statement/CommandStatement.java
@@ -8,9 +8,9 @@ public class CommandStatement {
         throw new IllegalStateException("Database class");
     }
 
-    public static void insert(PreparedStatement preparedStmt, int batchCount, int time, int user, int wid, int x, int y, int z, String message) {
+    public static void insert(PreparedStatement preparedStmt, int batchCount, long time, int user, int wid, int x, int y, int z, String message) {
         try {
-            preparedStmt.setInt(1, time);
+            preparedStmt.setLong(1, time);
             preparedStmt.setInt(2, user);
             preparedStmt.setInt(3, wid);
             preparedStmt.setInt(4, x);
